requirements.txt
fastapi==0.95.0
passlib==1.7.4
pydantic==1.10.6
PyJWT==2.6.0
pyodbc==4.0.35
python-dotenv==1.0.0
python_bcrypt==0.3.2
